% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_properties.R
\name{cdrProp}
\alias{cdrProp}
\title{AA properties calculation}
\usage{
cdrProp(.data, .prop = c("hydro", "polarity+turn"), .region = c("vhj",
  "h", "v", "j"))
}
\description{
AA properties calculation
}
