% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_theory.R
\name{entropy}
\alias{entropy}
\alias{kl_div}
\alias{js_div}
\alias{cross_entropy}
\title{Information measures}
\usage{
entropy(.data, .base = 2, .norm = F, .do.norm = NA,
  .laplace = 1e-12)
}
\description{
Information measures
}
