% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{get_rank}
\alias{get_rank}
\alias{set.rank}
\alias{set.index}
\title{Get ranks / indices for counts of clonotypes}
\usage{
get_rank(.data)
}
\arguments{
\item{.data}{Data frame or list with data frames.}

\item{.col}{Character vector with name of the column to use for ranking or indexing.}
}
\value{
Data frame with new column "Rank" or "Index" or list with such data frames.
}
\description{
Set new columns "Rank" and "Index":

set.rank <==> .data$Rank = rank(.data[, .col], ties.method = 'average')

set.index <==> .data$Index = 1:nrow(.data) in a sorted data frame by \code{.col}
}
