% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{group_from_metadata}
\alias{group_from_metadata}
\title{Get a character vector of samples' groups from the input metadata file}
\usage{
group_from_metadata(.by, .metadata, .sep = "; ")
}
\description{
Get a character vector of samples' groups from the input metadata file
}
