% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{lookup_clonotypes}
\alias{lookup_clonotypes}
\title{Search for clonotypes in external databases}
\usage{
lookup_clonotypes(.data, .db, .seq.col = NA, .v.col = NA,
  .db.name = NA)
}
\arguments{
\item{.data}{Input dataset.}

\item{.db}{Either data frame, data table or any other object with columns, or a path to the database file if ".db.name" is provided.}

\item{.seq.col}{Name of the CDR3 sequence column in the database.}

\item{.v.col}{Provide FALSE to match only by ".seq.col" if ".db.name" is provided.}

\item{.db.name}{Either NA or the name of the external database. Provide one of the supported database names to
skip specifying column names. Supported databases:

- "mcpas" - McPAS-TCR database, http://friedmanlab.weizmann.ac.il/McPAS-TCR/

- "vdjdb" - VDJDB database, https://vdjdb.cdr3.net/search}
}
\description{
Search for clonotypes in external databases
}
