% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{num_shared_clonotypes}
\alias{num_shared_clonotypes}
\title{Number of shared clonotypes.}
\usage{
num_shared_clonotypes(.data)
}
\arguments{
\item{.data}{List of repertoires of length 2 or more - either data frames, data tables or MonetDB tables.}

\item{.columns}{Character vector, which columns to use.}

\item{.norm}{Logical, if T than normalise the overlap value by multiplication of repertoires' sizes.}

\item{.head}{Overlap only .head clonotypes from each repertoire. -1 means overlap all clonotypes.}

\item{.verbose}{if T then output the progress.}
}
\value{
If length of .data is 2 than a single values, otherwise matrix with overlap values.
}
\description{
Number of shared clonotypes.
}
