% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectratyping.R
\name{spectratype}
\alias{spectratype}
\title{Analyse spectratype of immune repertoire}
\usage{
spectratype(.data, .quant = c("id", "count"), .col = c("nuc", "aa"),
  .gene = "v")
}
\arguments{
\item{.data}{The data to be processed. Can be \link{data.frame},
\link{data.table}, or a list of these objects.

Every object must have columns in the immunarch compatible format.
\link{immunarch_data_format}

Competent users may provide advanced data representations:
DBI database connections, Apache Spark DataFrame from \link{copy_to} or a list
of these objects. They are supported with the same limitations as basic objects.

Note: each connection must represent a separate repertoire.}

\item{.quant}{Select the column with data to evaluate}

\item{.col}{A string that specifies the column to be processed. Pass "nuc" for
nucleotide sequence or "aa" for amino acid sequence.}

\item{.gene}{A string with either "v", "j" or NA.}
}
\description{

}
