% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmers.R
\name{split_to_kmers}
\alias{split_to_kmers}
\alias{kmer_profile}
\title{Kmers processing.}
\usage{
split_to_kmers(.data, .k)
}
\arguments{
\item{.data}{Character vector or the output from \code{getKmers}.}

\item{.ic}{If T then return the position-weight matrix with self-information instead of frequencies.}
}
\description{
Kmers processing.
}
