% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{top_overlap}
\alias{top_overlap}
\title{Sequential overlap of top-by-count clonotypes.}
\usage{
top_overlap(.data, .fun, .step = 1000, ...)
}
\arguments{
\item{.data}{Either data frame, data table, MonetDB table or list of those, list of repertoires.}
}
\description{
Sequential overlap of top-by-count clonotypes.
}
