% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{trackClonotypes}
\alias{trackClonotypes}
\alias{tr.clo}
\alias{poisson_model}
\alias{norm_model}
\title{Clonotype tracking across different samples.}
\usage{
trackClonotypes(.data, .which = c(1, 15), .model = c("none", "poisson", "norm"),
.col = c("nuc", "aa"), .q = c(.025, .975), .sd = 1.05, .dup = c("merge", "remove"))

tr.clo(...)

poisson_model(.data, .q = c(.025, .975))

norm_model(.data, .q = c(.025, .975), .sd = 1.05)
}
\arguments{
\item{.data}{List of repertoires or a single repertoire.}

\item{.which}{Either 1) a vector of length two `c(X, Y)`, where `Y` is the number of the most abundant clonotypes to take from `X`'th data frame in the input list; 2) or a character vector of sequences to take from all data frames; 3) or a data frame with two / three columns - first for sequences, and second/third for gene segments.}

\item{.model}{Which model to use for estimation of confidence intervals for clonotypes: "none" for no model, "poisson" for poisson model, "norm" for normal model.}

\item{.col}{Column with sequences, either "nuc" for nucleotide sequences or "aa" for amino acid sequences.}

\item{.q}{Quantiles for models.}

\item{.sd}{Standart deviation for normal model.}

\item{.dup}{What to do with duplicated sequences - either merge and sum up their counts ("merge") or remove sequences with lower counts ("remove").}
}
\value{
Data frame with input sequences and counts / estimations for each of the input repertoires.
}
\description{
Track the dynamics of clonotypes across different repertoires.
}
\examples{
\dontrun{
immtr = trackClonotypes(immdata$data, "norm", c(3,100));
tidyr::drop_na(dcast(immtr, Sequence ~ Sample, value.var = c("Data")))
}
}
