% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{vis.immunr_hclust}
\alias{vis.immunr_hclust}
\alias{vis.immunr_kmeans}
\alias{immunr_dbscan}
\title{Visualisation of clustering results}
\usage{
vis.immunr_hclust(.data, .rect = F, .plot = c("clust", "best"), .return.grob = F)

vis.immunr_kmeans(.data, .point = T, .text = T, .ellipse = T,
.point.size = 2, .text.size = 10, .plot = c("clust", "best"), .return.grob = F)

vis.immunr_dbscan(.data, .point = T, .text = T, .ellipse = T,
.point.size = 2, .text.size = 10, .plot = c("clust", "best"))
}
\arguments{
\item{.data}{Clustering results from \link{repOverlapAnalysis} or \link{geneUsageAnalysis}.}

\item{.rect}{Passed to \link{fviz_dend} - whether to add a rectangle around groups.}

\item{.point}{If TRUE then plot sample points. Passed to \link{fviz_cluster}.}

\item{.text}{If TRUE then plot text labels. Passed to \link{fviz_cluster}.}

\item{.ellipse}{If TRUE then plot ellipses around all samples. Passed to "ellipse" from \link{fviz_cluster}.}

\item{.point.size}{Size of points, passed to "pointsize" from \link{fviz_cluster}.}

\item{.text.size}{Size of text labels, passed to labelsize from \link{fviz_cluster}.}

\item{.plot}{A character vector of length one or two specifying which plots to visualise.
If "clust" than plot only the clustering. If "best" than plot the number of optimal clusters.
If both than plot both.}

\item{.return.grob}{If TRUE then return grob instead of plot.}
}
\description{
Visualisation of clustering results
}
\seealso{
\link{vis}, \link{repOverlapAnalysis}, \link{geneUsageAnalysis}
}
