% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{vis.immunr_mds}
\alias{vis.immunr_mds}
\alias{vis.immunr_pca}
\alias{vis.immunr_tsne}
\title{PCA / MDS / tSNE visualisation (mainly overlap / gene usage)}
\arguments{
\item{.data}{Output from \code{immunr_pca} or \code{immunr_mds}.}
}
\description{
PCA / MDS / tSNE visualisation (mainly overlap / gene usage)
}
