% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{vis.immunr_public_repertoire}
\alias{vis.immunr_public_repertoire}
\title{Public repertoire visualisation}
\arguments{
\item{.data}{Public repertoire data — an output from the \link{pubRep} function.}

\item{.x.rep}{Either indices of samples or character vector of sample names
for the x-axis. Must be of the same length as ".y.rep".}

\item{.y.rep}{Either indices of samples or character vector of sample names
for the y-axis. Must be of the same length as ".x.rep".}

\item{.title}{The text for the title of the plot.}

\item{.ncol}{An integer number of columns to print in the grid of pairs of repertoires.}

\item{.point.size.modif}{An integer value that is a modifier of the point size.
The larger the number, the larger the points.}

\item{.cut.axes}{If TRUE then axes limits become shorter.}

\item{.density}{If TRUE then displays density plot for distributions of clonotypes
for each sample. If FALSE then removes density plot from the visualisation.}

\item{.lm}{If TRUE then fit a linear model and displays an R adjusted coefficient
that shows how similar samples are in terms of shared clonotypes.}

\item{.radj.size}{An integer value, that defines the size of the The text
for the R adjusted coefficient.}

\item{.return.grob}{If TRUE then returh the gridArrange grob instead of the plot.}
}
\description{
Visualise public clonotypes.
}
\seealso{
\link{pubRep}
}
