% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{vis_box}
\alias{vis_box}
\title{Visualisation of distributions using boxplots}
\arguments{
\item{.data}{Input matrix or data frame.}

\item{.by}{Pass NA if you want to plot samples without grouping.

You can pass a character vector with one or several column names from ".meta"
to group your data before plotting. In this case you should provide ".meta".

You can pass a character vector that exactly matches the number of samples in
your data, each value should correspond to a sample's property. It will be used
to group data based on the values provided. Note that in this case you should
pass NA to ".meta".}

\item{.meta}{A metadata object. An R dataframe with sample names and their properties,
such as, age, serostatus or hla.}

\item{.title}{The text for the title of the plot.}

\item{.ncol}{Number of columns to display. Provide NA (by default) if you want the function
to automatically detect the optimal number of columns.}

\item{.labs}{Character vector of length two with names for x-axis and y-axis, respectively.}

\item{.melt}{If TRUE then apply \link{melt} to the ".data" before plotting.
In this case ".data" is supposed to be a data frame with the first character column reserved
for names of genes and other numeric columns reserved to counts or frequencies of genes.
Each numeric column should be associated with a specific repertoire sample.}

\item{.points}{A logical value defining whether points will be visualised or not.}

\item{.test}{A logical vector whether statistical tests should be applied. See "Details" for more information.}

\item{.signif.label.size}{An integer value defining the size of text for p-value.}

\item{.title}{The The text for the plot's title.}

\item{.leg.title}{The The text for the plots's legend. Provide NULL to remove the legend's title completely.}

\item{.legend}{If TRUE then displays a legend, otherwise removes legend from the plot.}
}
\description{
Visualisation of distributions using ggplot2-based boxplots.
}
\seealso{
\link{vis.immunr_gene_usage}, \link{geneUsage}
}
