% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{vis_circos}
\alias{vis_circos}
\title{Visualisation of matrices using circos plots}
\arguments{
\item{.data}{Input matrix.}

\item{.title}{The The text for the title of the plot.}

\item{...}{Other arguments passed to \link{chordDiagram} from the 'circlize' package.}
}
\description{
Visualise matrices with the \link{chordDiagram} function
from the circlize package.
}
\seealso{
\link{vis}, \link{repOverlap}.
}
