% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{vis_logo}
\alias{vis_logo}
\title{Logo - plots for amino acid and nucletide profiles.}
\arguments{
\item{.data}{Output from the \code{kmer.profile} function.}

\item{.replace.zero.with.na}{if T then replace all zeros with NAs, therefore letters with
zero frequency wont appear at the plot.}

\item{.jitter.width, .jitter.height, .dodge.width}{Parameters to \code{position_jitterdodge}
for aligning text labels of letters.}
}
\value{
ggplot2 object
}
\description{
Plot logo-like graphs for visualising of nucleotide or amino acid motif sequences / profiles.
}
\examples{
\dontrun{
d <- kmer_profile(c('CASLL', 'CASSQ', 'CASGL'))
vis_logo(d)
}
}
